requirements.txt
music-recommender==1.0.2
requests